src/pyproject.toml
[project]
name = "flask-api"
version = "4.0.0"
description = "API Application with Flask for CRUD users"
authors = [
    {name = "Rubén Rodríguez Ramírez", email = "rubenrod.018@gmail.com"}
]
license = {text = "MIT License"}
readme = "README.md"
requires-python = ">=3.13,<4"
dependencies = [
    "python-dotenv (>=1.1.0,<2.0.0)",
    "faker (>=37.3.0,<38.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "mysqlclient (>=2.2.7,<3.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "celery (>=5.5.2,<6.0.0)",
    "flower (>=2.0.1,<3.0.0)",
    "python-docx (>=1.1.2,<2.0.0)",
    "python-magic (>=0.4.27,<0.5.0)",
    "xlsxwriter (>=3.2.3,<4.0.0)",
    "flask (>=3.1.1,<4.0.0)",
    "flask-mail (>=0.10.0,<0.11.0)",
    "flask-marshmallow (>=1.3.0,<2.0.0)",
    "flask-restx (>=1.3.0,<2.0.0)",
    "flask-security-too (>=5.6.2,<6.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "flask-migrate (>=4.1.0,<5.0.0)",
    "flask-jwt-extended (>=4.7.1,<5.0.0)",
    "marshmallow-sqlalchemy (>=1.4.2,<2.0.0)",
    "dependency-injector (>=4.47.0,<5.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "google-api-python-client (>=2.173.0,<3.0.0)",
    "google-auth-httplib2 (>=0.2.0,<0.3.0)",
    "google-auth-oauthlib (>=1.2.2,<2.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
coverage = "^7.9.1"
freezegun = "^1.5.2"
pytest = "^8.4.1"
pytest-xdist = "^3.7.0"
sqlalchemy-utils = "^0.41.2"
pre-commit = "^4.2.0"
pylint = "^3.3.7"

[tool.poetry]
package-mode = false
